+++ client-app/shared/layout/components/header/_internal/mobile-menu/menus/multi-organisation-menu.vue
@@ -50,19 +50,15 @@
 
 <script setup lang="ts">
 import { ref, useTemplateRef, computed } from "vue";
-import { useAuth, useErrorsTranslator } from "@/core/composables";
-import { IdentityErrors } from "@/core/enums";
 import { useUser, useUserOrganizations } from "@/shared/account";
-import type { IdentityErrorType } from "@/core/api/graphql/types";
+import { useOrganizationSwitcher } from "@/shared/account";
 
 const { organizations, hasNextPage, loading, pagesCount, currentPage, loadOrganizations } = useUserOrganizations();
-const { user, isMultiOrganization, switchOrganization, organization } = useUser();
-const { errors: authErrors } = useAuth();
-const { translate } = useErrorsTranslator<IdentityErrorType>("shared.account.sign_in_form.errors");
+const { user, isMultiOrganization, organization } = useUser();
+const { switchError, trySwitch } = useOrganizationSwitcher();
 
 const contactOrganizationId = ref(user.value?.contact?.organizationId);
 const scrollContainer = useTemplateRef("scrollContainer");
-const switchError = ref<string | null>(null);
 
 const organizationsWithoutCurrent = computed(() =>
   organizations.value.filter((item) => item.id !== organization.value?.id),
@@ -73,14 +69,9 @@ async function selectOrganization(): Promise<void> {
     return;
   }
 
-  switchError.value = null;
-  const succeeded = await switchOrganization(contactOrganizationId.value);
+  const succeeded = await trySwitch(contactOrganizationId.value);
 
-  if (!succeeded && authErrors.value?.length) {
-    const lockedError = authErrors.value.find(
-      (e) => e.code === IdentityErrors.USER_IS_LOCKED_IN_ORGANIZATION || e.code === IdentityErrors.USER_IS_LOCKED_OUT,
-    );
-    switchError.value = (lockedError ? translate(lockedError) : translate(authErrors.value[0])) ?? null;
+  if (!succeeded) {
     contactOrganizationId.value = user.value?.contact?.organizationId;
   }
 }
